requirements.txt
Django==4.0
djangorestframework==3.13.1
mysqlclient==2.1.0
ddtrace==0.56.1
celery==5.2.1
requests==2.26.0
